# src/AppForSEII2526.Web/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Receipt/GetReceiptDetail": {
      "get": {
        "tags": [
          "Receipt"
        ],
        "operationId": "GetReceiptDetail",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDetailDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Receipt/CreateReceipt": {
      "post": {
        "tags": [
          "Receipt"
        ],
        "operationId": "CreateReceipt",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptForCreate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptForCreate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptForCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ReceiptDetailDTO": {
        "required": [
          "applicationUserName",
          "applicationUserSurname",
          "deliveryAddress",
          "receiptItems"
        ],
        "type": "object",
        "properties": {
          "applicationUserName": {
            "minLength": 1,
            "type": "string"
          },
          "applicationUserSurname": {
            "minLength": 1,
            "type": "string"
          },
          "deliveryAddress": {
            "minLength": 1,
            "type": "string"
          },
          "receiptItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReceiptItemDTO"
            }
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "dateOfReceipt": {
            "type": "string",
            "format": "date-time"
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ReceiptForCreate": {
        "required": [
          "applicationUserName",
          "applicationUserSurname",
          "deliveryAddress",
          "receiptItems"
        ],
        "type": "object",
        "properties": {
          "applicationUserName": {
            "minLength": 1,
            "type": "string"
          },
          "applicationUserSurname": {
            "minLength": 1,
            "type": "string"
          },
          "deliveryAddress": {
            "minLength": 1,
            "type": "string"
          },
          "receiptItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReceiptItemDTO"
            }
          }
        },
        "additionalProperties": false
      },
      "ReceiptItemDTO": {
        "required": [
          "cost",
          "model",
          "name",
          "scale"
        ],
        "type": "object",
        "properties": {
          "repairID": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "scale": {
            "minLength": 1,
            "type": "string"
          },
          "cost": {
            "type": "number",
            "format": "double"
          },
          "model": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}